backend/pyproject.toml
[project]
name = "conversation-persistence"
version = "0.1.0"
description = "Stateless conversation persistence API for AI-powered Todo Application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "sqlmodel>=0.0.14",
    "pydantic>=2.5.0",
    "uvicorn[standard]>=0.27.0",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.1.0",
    "mcp>=1.25,<2",
    "openai-agents>=0.0.3",
    "google-genai>=1.56.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<5",
    "email-validator>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "ruff>=0.2.0",
    "aiosqlite>=0.19.0",
    "pyyaml>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
